:art: Refactor: 프로필 메뉴 컴포넌트 생성 & 파일 구조 변경
